+++ main.h
@@ -78,8 +78,8 @@ class Edge_property
 struct Node
 {
     uint64_t id;
-    float lon;
-    float lat;
+    double lon;
+    double lat;
     ushort uses;
 
     Node() : uses(0) {};
